requirements.txt
azure-functions
requests
azure-iot-device
google-api-python-client